move hash method from handle_misc to handle_files
